requirements.txt
pandas>=1.5.0
scikit-learn>=1.1.0
python-Levenshtein>=0.12.0
transformers>=4.35.0
tqdm>=4.64.0
matplotlib>=3.6.0